ref: move strategies in their own directory
