+++ src/main/java/fr/uga/l3miage/pc/strategies/DonnantDonnantAleatoire.java
@@ -1,7 +1,6 @@
-package fr.uga.l3miage.pc.models.strategies;
+package fr.uga.l3miage.pc.strategies;
 
 import fr.uga.l3miage.pc.enums.EnumIdJoueur;
-import fr.uga.l3miage.pc.interfaces.SimpleStrategy;
 import fr.uga.l3miage.pc.models.Tour;
 
 import java.util.Random;
